Refactor processing queue posting a LOT, also make it so actions have a virtual dtor
Need to still fix the memleak that happens sometimes.

This commit also syncs the branch to using the 1.2.10 database format,
so the Beast server can be interchanged with 1.2.10 and Beast.

+++ src/Database/Config.h
@@ -124,7 +124,7 @@ struct Config
 	/**
 	 * Moderator permissions.
 	 */
-	uint8_t ModPerms;
+	uint16_t ModPerms;
 
 
 	//FileMode FileUploadMode;
